Unify GPIO pin sentinel to 0xFF across device configs, add missing validation
Audits every GPIO/channel field across device configs (see new
docs/pin-configuration-conventions.md) and unifies the "not configured"
sentinel to uint8_t 255, replacing a mix of ad hoc conventions: real-looking
but arbitrary defaults (gpio_switch=2, binary_sensor/onewire_bus=4,
analog_port_input/cd74hc4067.sigPin=34, dht11=17, ds1302=25/26/27), and
per-family negative-int sentinels (spi_bus.misoPin int16_t=-1,
st7735.resetPin int8_t=-1).

Fields genuinely optional in hardware (hd44780.backlightPin,
cd74hc4067.enablePin) and fields with a real hardware-standard default
(i2c_bus, spi_bus.sckPin/mosiPin, the I2C-backpack lcd1602/lcd2004 channel
mapping) are left untouched.

Two fields needed a config version bump because their type/size changed
(spi_bus.misoPin V1->V2, st7735.resetPin+missing chipSelectPin/dcPin
validation V5->V6); everything else is an in-place default-literal edit,
since defaults aren't part of a struct's persisted bytes and every real
create/update request already sends an explicit value.

Also fixes two real bugs found during the audit: hd44780's six direct-wire
pins used to all default to 0, so a freshly created device failed its own
"pins must be distinct" check; onewire_bus.gpioPin had no range validation
at all.

+++ src/devices/bus/spi/ISpiBusDriver.h
@@ -17,7 +17,7 @@ public:
     ISpiBusDriver& operator=(ISpiBusDriver&&) = delete;
     virtual ~ISpiBusDriver() = default;
 
-    virtual bool begin(uint8_t host, uint8_t sckPin, uint8_t mosiPin, int16_t misoPin) = 0;
+    virtual bool begin(uint8_t host, uint8_t sckPin, uint8_t mosiPin, uint8_t misoPin) = 0;
     virtual bool end() = 0;
     virtual bool beginTransaction(uint32_t clockHz, uint8_t dataMode, uint8_t bitOrder) = 0;
     virtual void endTransaction() = 0;
